Log no save ui reason if no dataset set.
The TYPE_CONTEXT_COMMITED will not be log if there is no dataset in
the FillResponse currently. We add no save ui reason in the event
history that helps AutofillService to know the reason the save ui
not shown. Because the above reason, the event will not be logged
even the no save ui reason is set. But it's still useful to log the
event to allow the AutofillService to know the no save ui reason so
we remove the limitation here.

Bug: 217584507
Test: atest InlineFillEventHistoryTest
Test: atest FillEventHistoryTest

Change-Id: I523c373c9ae12230c27ede81b0ac88d30ee5c27f

Former-commit-id: 69a1dd6a74dc5d82d2d208fe7b43246c42fe5732