Partially undo "Try to avoid overlapping job execution."
This reverts the behavior change in
commit 84ab42933fe56e0a5502ad3fac2bac07899f0de4 [formerly 61ffd79f39d1f68b452f503e2f403aa69d9ff6d8].

However, we log the instances where jobs are started while a similar one
is already running. There can be legitimate instances when there's an
overlap, so no reason to log as a WTF. Limit logging to when DEBUG is
enabled to avoid a performance hit on end-user devices.

Some documentation and logging additions were retained.

Reason for revert: Not necessary.

Bug: 231148615
Test: N/A
Change-Id: I5a89ad3dbd5c5e94b262485e8a418086ca862e9b

Former-commit-id: e02c09db3f98119d69ed37ce617aa7188e83b8a9