Don't blend front scrim on AOD
The scrim tint is always blended with the main color, as determined by
color extraction and the theme. For AOD, it should always fade to dark
and never be blended with a light color. This is most noticeable in
low light environments when the front scrim is animated and then
blended with white.

Fixes: 223592183
Test: atest ScrimControllerTest + manual (phone in low light,
transition to AOD)

Change-Id: I6b7b78b26bcbf63989580d2f46981d79bcc9a99b

Former-commit-id: c60ff20ecb0caec9daf4e23398b9b1cdac01d7a6