Fix duplicate-call issues in remote-display-change
The display-change code was just accumulating callbacks without
ever removing them. This is both a memory leak and was calling
stale callbacks which yielded out-of-order changes.

Additionally, it was calling all callbacks for each individual
callback received (meaning the first of 2 in-flight remotes
would call both callbacks before the second was received). This
meant that the transactions were being used in the wrong
callbacks and that some callbacks would be called out-of-order.

Lastly, the display-rotation failure-case was wrong. Its supposed
to fall-back on the new rotation not the original rotation.

Bug: 232477270
Test: atest PinnedStackTests -- shouldn't crash due to
      rotation-outside-of-transition anymore.
Change-Id: Icc7a90448edb648a8110cd83018a1477111beedd

Former-commit-id: 271f4fca0100908be04a56f1c8eba63103ceb8dd