Defer sending config change to cached process
Though there was commit e839c2c to defer on client side, if
CachedAppOptimizer decides to freeze the cached process, the
async binder buffer may still be filled, and may lead to
binder transaction fail right after unfreezing because the
free buffer is not enough.

So now there is no IPC for config change of cached process,
and will send the latest change to the client when the process
state is changed from cached to non-cached.

Bug: 217934273
Test: WindowProcessControllerTests#testCachedStateConfigurationChange

Change-Id: I09c2feffba2fc124c623796af69de9f6adb60a39

Former-commit-id: c0537b750d0f29dab4a792e4d9d728d7ef69354e