Create proper ui context if original context is not a ui context
Since Android API 31 there is a new check in StrictMode, this
check ensures that any ui-related operations involving context
object are only performed on the ui context. If the context is
not the ui context, it will fail with error when calling the
ui-related methods.

So we need to create proper ui context if original context is
not a ui context.

Bug: 220698651
Test: atest CtsVoiceInteractionTestCases
Test: manual. Test VoiceInteractionSession UI.
Change-Id: Ibbeae9b135b90078b3b9e36a2d0c5a92f6184b2a

Former-commit-id: c6931d97cb20f569e86d8624e0cae49496423cb7