Update vis of face acquisition + err messages
When fingerprint is also enrolled, don't surface
FACE_TIMEOUT errors or face acquisition messages.

This CL also adds the ability to override the face
acquisition messages that can show when fp is also
enrolled. This can be done via adb. For example,
the following command would enable
    FACE_ACQUIRED_MOUTH_COVERING_DETECTED and
    FACE_ACQUIRED_DARK_GLASSES_DETECTED
 adb shell settings put global coex_face_help_msgs "26\|25"
This override is for temporary teamfooding and should be removed
before release.

Test: manual
Test: atest KeyguardIndicationControllerTest
Bug: 231733975
Change-Id: I1fb569adcf91576f89aa1fb068db8bdd030ea645

Former-commit-id: 8548493bdf5f1685fe30e7331d14aaa8c358b232