Fix screen off after panel expansion
While unlocked, when the panel was expanded the
DozeParameters#shouldControlScreenOff was not properly getting reset
after the panel was collapsed. This was leading to overlapping
animations as the panel state changed. The panel change event that
DozeParameters cares about originates in NotificationPanelViewController
so make sure the check in UnlockedScreenOffAnimation uses this same
field, and not the one in PanelViewController, which may have a
different value.

Fixes: 203644049
Test: manual (unlock/lock over and over and over ...)
Change-Id: Iaa357eed4afae53f8f96995c570b5df1a83c8123

Former-commit-id: bd09f730bac4e47cd14585966d536ab0a89231f5