Don't inflate mSplitDecorManager in sync transaction callback
The current code inflates the UI in the sync transaction callback. This
can result in a deadlock since inflate could block waiting on RT. The RT
could be blocked waiting for the sync transaction to get applied to free
up a buffer. But since the sync transaction won't get applied until after
the inflate, it would deadlock.

Test: Not stuck in split
Fixes: 231929473
Change-Id: I751aa2d68ec96f7f77ba4145d810cccf764075ea

Former-commit-id: 04efcbae292f59ab6f66f5126d2e3416486d5161