SystemService.isUserSupported uses isProfile
Previously, SystemService's isUserSupported() used isManagedProfile, but
in many cases, it makes more sense to use isProfile:
most of the particular cases already done probably just care about profile,
not managed profile.

So we introduce an isProfile(). We also modify how the userType is
stored in this class, to make it more extensible in the future.

Bug: 170249807
Test: treehugger. This is a no-op since there currently aren't really
other supported profile types anyway

Change-Id: I43168939001d0e68ed6de95127161f1197683b10

Former-commit-id: e554479c028dd0c75ff9cad5f0a2f21d1e4f3019