Support to launch the matching task into a closer adjacent target task
When launching trampoline activities into split screen while having the
final activity in another side, it'll move the matching task, the task
containing the final activity, to another side of the split and result
to split dismissed due to one side of the split is empty.

To prevent split screen got dismissed during trampoline launch, adjust
to apply a closer adjacent launch target to the matching task.

Bug: 224901460
Test: atest ActivityStarterTests
Test: put a trampoline launched activity into split,
      drag and drop the same app icon into another side of the split,
      verified it won't dismiss split screen.
Change-Id: I17d401a6648214fbc21c5fb1799267b34abdb355

Former-commit-id: deaad42323ea79a677c810f4ddbb22d1abf38768