[LockScreen] Update bar state in NPVC.
When view is attached to window in NPVC, we add a listener to the bar
state controller; however we do not update to the existing bar state if
any. This means that if the listener is added before the bar state
change is made, then we will default to 0 (SHADE) for NVPC. We have
experienced strange LS states because of this.

Bug: 230911766
Test: Added a 5s delay to adding the callback to experience the weird
state. Added a unit test.

Change-Id: Ieb6d6537eb002b03beceef6f701b826df48feeeb

Former-commit-id: 036054e88c356290a14ddccc62d55fff4ba719ed