Reset original task surface after reparenting to PiP
Currently if the shell transitions are disabled (default), the auto-enter
pip will trigger an animation via Recents, animating the original task
surface and populating the mLastRecentsAnimationTransaction. As we
reparent that activity into the pip window, we also reset the state of the
original task surface in clearLastRecentsAnimationTransaction.

When shell transitions are enabled, the auto-enter pip does not animate via the same mechanism and the mLastRecentsAnimationTransaction is null. We still animate the original task surface but nothing resets it after the
activity is being reparented.

This CL fixes that second scenario by applying a reset transform to the
original task surface even if there was no recents animation transaction.

Bug: 232981505
Test: manual, open ApiDemos, autoenter pip, return to full screen via pip menu

Change-Id: Ibc42c1277e045bf5042678340b9d68428cb2e626

Former-commit-id: b464667b771b1d976359e7cd8b83f52d6245774c