Adding a reserve of temporary quota for alarms
Adding a mechanism to allocate temporary per-app quotas that can be used
by them after running out of their standard standby quota.
This can be used to give back apps quota in case they are not directly
being opened by the user but interact with the user in indirect ways
like notifications.
Currently this quota is set to expire 24 hours from the time it is last
replenished.

Test: atest FrameworksMockingServicesTests:AlarmManagerServiceTest
Existing CTS tests pass:
atest CtsAlarmManagerTestCases

Bug: 231925670
Change-Id: I18e5b2fbbb40b378b2853c4e86fa4729f87f76e4

Former-commit-id: 355586523305bba2246c76aa5954bb48f073049b