Define the SystemUI typography
This CL defines the material typography that is going to be used by
Compose code in SystemUI. Most of the code from this CL was forked from
the Google Material implementation in google3 [1, 2, 3, 4], so that our
typography is consistent from 1P apps (and with the Google Material
spec). For this reason, I changed only what I had to + reformatted the
files using ktfmt. The direct references to "google-sans" will be
removed in http://ag/19500381 to keep this CL as close as possible from
the google3 classes.

Note that this new typography might not be 100% consistent with the
current text styles used in SystemUI, given that they are not consistent
with the new M3 spec. Those kind of inconsistencies will be fixed later,
as we figure out how SystemUI is going to be aligned with M3 in general
(and not w.r.t. Compose in particular).

[1] http://google3/java/com/google/android/libraries/material/compose/Typography.kt
[2] http://google3/java/com/google/android/libraries/material/compose/tools/tokens/library/md/first_party/TypographyTokens.kt
[3] http://google3/java/com/google/android/libraries/material/compose/tools/tokens/library/md/first_party/TypeScaleTokens.kt
[4] http://google3/java/com/google/android/libraries/material/compose/tools/tokens/library/md/first_party/asynctype/TypefaceTokens.kt

Bug: 230605885
Test: Manual
Change-Id: I116ef944c79e09fb0d4b0914c3da71a946fc169a

Former-commit-id: 8c222cabb631082f21abfa3abcd56c5ec62aee9a