Run unlock animation on MODE_ONLY_WAKE
When the screen times out, the device may not immediately lock based
upon a user setting. In this case, keyguard is visible but not fully
locked. If the user uses bio auth, MODE_ONLY_WAKE is used but the
keyguard is not properly dismissed. Issue a call to keyguardDone so
that the correct animation is shown.

Fixes: 231553219
Test: atest BiometricsUnlockControllerTest

Change-Id: Ib3fa05f2a1054b122a80ca5d27b0a99d12dd27f2

Former-commit-id: 5481291dca0fa77dee70ec097d305130181facb1