Revert "Rebind ContentCaptureService when the binderDied."
This reverts commit 0ff6499c59553a37452ade19a39472626f27b7a5 [formerly 4fc5c86bb9945be49cd77140facf62ba70b5b7e4].

We introduce a rebind mechanism to rebind died ContentCaptureService
because the apps tht has dependency with ContentCaptureService. But
it looks like the rate is raised for app to get the null content
capture manager.

Because the underlying issue is fixed, we don't need this workaround
now. We can also monitor if the null content capture sympton is
resolved.

Bug: 227353916
Test: manual. Add dependency and update the app, the ContentCapture
can be rebind without the this fix.
Reason for revert: To monitor if the null content capture issue
caused by this change.

Change-Id: Ibe5d6e2b5f1434243b655289fdc60174ff33a5c0

Former-commit-id: fa414929e1a618325a2c4ad3389eae3706564efd