Use System Property to Control Animator Pausing
This change sets the default behavior of animator pausing by checking if
a system property is set and useing that value if present else it
defaults to true. This should allow the use of build properties to be
able to toggle the behavior.

Bug: 233391022

Test: Manual, adding logging to verify the values being set by the call
      to the sysprop library.

Ignore-AOSP-First: Testing internal change, will cherry pick to AOSP.
Change-Id: Ib1b34585c564bf4f310441c6856412a798022900

Former-commit-id: 063ae10a70aa27438f275e652e877897037aec83