Change TimeDetectorStrategy.Environment API
Switch TimeDetectorStrategy.Environment to returning
ConfigurationInternal in place of an individual property.

This is in preparation for implementing the methods to update
time detector configuration and support listeners. More properties will
be added to ConfigurationInternal / removed from Environment in later
commits.

This mirrors the situation in com.android.server.timezonedetector and is
part of a sequence of changes to make com.android.server.timedetector
follow the patterns for configuration used there.

Bug: 172891783
Bug: 229740080
Test: atest com.android.server.timedetector android.app.time
Change-Id: I0ffe467ede4b6b0279b284108a368c5265a323a0

Former-commit-id: 8247153b486838872d18d7c61215b5d03d3f3f92