Fork SystemUiController from Accompanist
This CL forks the SystemUiController class from Accompanist [1] to
easily change the color of the system bars and of the status bars icons.
See the documentation in [2]. This class makes it really easy to create
Activities with system bars that behave as expected without having to
touch a single file of XML, as shown in go/sysui-compose-docs#activity.

Given that Accompanist is an incubator for experimental Compose features
before they are upgraded to androidx, and given that this is a simple
class, I think that forking it is a better idea than trying to set up
the prebuilt dependencies on the library. Hopefully this will be part of
the androidx libraries soonish.

[1] https://github.com/google/accompanist/tree/main/systemuicontroller/src/main/java/com/google/accompanist/systemuicontroller
[2] https://google.github.io/accompanist/systemuicontroller/

Test: Manual
Bug: 230605885
Change-Id: Idd142b33eb326f9ed8ce4ca9362390155f35e238

Former-commit-id: 162fefab20d9bed663470528362ef1115ad5e4ae