Only consider occluding transient below-task for wallpaper target
Otherwise if keyguard is not locked (recents will be moved to top),
the wallpaper target may be updated to the window of below-task
temporally when swiping-up. Which causes the wallpaper window cannot
apply the same fixed-rotation transform if the recents activity has
the transform. Such as when swiping up to portrait home from a
landscape activity, the wallpaper will be rotated to landscape and
trigger an unnecessary display rotation animation.

Bug: 230582311
Test: atest WallpaperControllerTests# \
            testFixedRotationRecentsAnimatingTask
Test: adb shell setprop persist.wm.debug.shell_transit 1; reboot
      Swipe landscape app to home and no additional display
      rotation animation for rotated wallpaper.
      With show-when-locked activity on lockscreen, the wallpaper
      can be visible when swiping up.

Change-Id: Ibd47e9d78c2f6767a78255db62284bbbf176918c

Former-commit-id: da3433f0d9ec2138ef4ef35e7590f8832355653c