Fixing empty split shade after rotation from expanded QS on lockscreen
In portrait, swipe from the top of the screen goes to expaned QS but state of StatusBar is KEYGUARD. That makes split shade confused, alphas and translations are all over the place.
The solution is to force change state to SHADE_LOCKED which makes more sense in that case in split shade.

Also fixing comments of StatusBarState.

Fixes: 232396574
Test: Go to portrait shade -> go to expanded QS with swipe from the top -> rotate to split shade and see everything displayed correctly
Test: NotificationPanelViewControllerTest
Change-Id: Ib4fe90a7b784cfe755179518c6e19e8fd7c10436

Former-commit-id: cd49a28e4f38aa70e5a26aa804c71051e260b695