Always set the top roundness of child rows to 0.
The bug demonstrated that only the top radius was getting into a dirty state.  When I saw this function was resetting the bottom roundness and never touching the top roundness, it was clear that we just needed to set the top roundness.  Even though it's hard to reproduce the issue where top roundness is incorrectly huge, it's trivial to reproduce the issue where the top roundness is incorrectly small, and this fix definitively addresses that state.

Test: post auto-groupable Heads Up notifications.  Notice after grouping that child rows now never have any internal corner radii.
Fixes: 227433308
Change-Id: I3cd1aa23ad8333f35ec837edd39bc12aeb1573b6

Former-commit-id: ef4782f55d7c00493fe6cd9e10c2a082c2a819e3