Limit preview stabilization mandatory stream sizes to 1440p (instead of RECORD).
Also fix a logic error in an if check while retrieving mandatory streams
for preview stabilization. This was not caught by CTS since for devices
where preview stabilization was supported, skipped the test. A
modification has been made to CTS RobustnessTest to ensure that the
preview stabilization support and mandatory streams are in agreement
with each other.

Bug: 228245412

Test: atest RobustnessTest.java on device supporting preview stabilization

Change-Id: I18c7f3c728862866b8f088da3afe222afbad7051
Signed-off-by: Jayant Chowdhary <jchowdhary@google.com>

Former-commit-id: d83e9b5325bcc4d771b3f4f7afda48a5c1b37d7f