Grab pointer into array post-resize()
This is a defensive change.  The current code is safe due to the
resize() which happens 30 lines above this one.  However, if that
resize() were to change or disappear, then this code would be
dangerous, as the resize() could potentially cause a reallocation,
and move our memory buffer.

Since it's no additional cost to grab the pointer into the
array after the resize(), we do that.

Bug: 232940948
Test: TreeHugger
Change-Id: I29b6cbeb064c7654eb21d2e42e05a0587604c32b

Former-commit-id: 851c62d45c9b26d73d50430ace563b40e5adce3f