Support Activity launches above the expandable
This CL adds support for activity launches where the opening window is
opening above the expandable surface.

Before this CL, we would automatically assume that the opening window
would be below the expandable surface, which is currently always the
case given that the animator is used only in the shade and status bar at
the moment (and those are above any app that we open).

After this CL, the ActivityLaunchAnimator has an additional parameter
for launches that happen above the expandable surface, which happen when
launching an Activity from a normal app or from a launcher.

Bug: 235576267
Test: Manual
Change-Id: I49c6095ffe52119d117af6c597463542b4630bbd

Former-commit-id: 7bb867220fd86ca517b393030b61cfbc1dab3e36