Set occluded = false if occlude remote animation is cancelled.
Due to a bug, this was previously not needed (and could
cause issues) since WM would call setOccluded(false)
directly if the animation was cancelled. However, that
was due to a flag mismatch bug fixed in ag/17776445.

With that bug fixed, WM is never expected to call
setOccluded directly, so we need to do it when the
animation is cancelled.

Fixes: 232002936
Fixes: 232296784
Test: atest SystemUITests
Test: launch/kill occluding activities
Change-Id: I9eadd8940f1e0d71c6b627ccc0301284dfefbc2b

Former-commit-id: ec8486238bae6deaecc1a829f4095c1b768e8e20