Add MotionEvent information to exception
When indices are validated and an invalid index is encountered,
currently there is a Java exception raised.

The caller may be able to catch that exception in Java. However, we are
also finding crashes in the wild where these exceptions are unchecked.

As a first pass, let's simply add event information to understand
exactly how this could happen.

Bug: 233163975
Test: atest libinput_tests inputflinger_tests
Change-Id: If717b9ca90d85338752a0f911d6fb66bbbf4fe4a

Former-commit-id: 2885a1c490c531b8686a6e17e3f28ee5fa34d66a