Remove lowpan internal API
This API is not used: marked as @hide with a commented out @SystemApi;
the AIDL server is not implemented in AOSP nor internally. There is one
extra client from the SEPolicy: wpantund; but there is no build rule for
this binary, and no reference for it in AOSP. This API appears be an
artifact from the initial import of Android Things back in 2017.

I confirmed with the previous owners that there is no intention to use
this API in the future (see bug).

This changes reduces the size of framework.jar by 77kb on
bramble-userdebug.

Bug: 235018188
Test: TH
Change-Id: Id66f5ed23ed26c7185f36454f168e34997be3f8f
Merged-In: Id66f5ed23ed26c7185f36454f168e34997be3f8f
(cherry picked from commit 27df6595ae2e68d33293019787f3a454deae990e [formerly 7640656825ede6ff7a90ccea9593394179fcc218])

Former-commit-id: 231574ce345ae6dc89a72de4d307aa754565a8ab