Fix app links verification resetting user state
In cases where domain verification comes in and does not take over
verification from another app (the owning app has link handling
disabled or the domain was already previous verified), the user
selection for that domain should not be reset.

This preserves user approved apps for links which the user has
disabled the official verified app when that verified app update,
as this event kicks off a verification through StatementService,
which may have caused the reset with the buggy code.

Bug: 221059351

Test: atest DomainVerificationPackageTest
Test: atest DomainVerificationIntentStandaloneTests
Test: manual, reset verification using shell set-app-links
Test: manual, reset verification using verified app update

Change-Id: Ia8aa70be485b2f94920fd97275fd16178895ef33

Former-commit-id: 5697a4c37eb7e49d3c2d7292f3486b6407f21a44