Fix yet another clipTopAmount usage
Just like I3dec0e99111777ca37eddf79a33487fc0c8dc99c and I01293c0aecbeee559e97f348ff7c4151a64538ab, this imposes a minimum of 0 when using the clipTopAmount to determine touches.  I had done this for group children and a11y rectangles, but I missed the NSSL's calculation of which child is touched.

This particular method is only called when gestures are being intercepted, such as drag, and because it looks at children in increasing index rather than decreasing (as the platform touch handling does), the bug doesn't present when you drag on a different notification, it only presents if you drag somewhere that has no notificaton.

Change-Id: I2d41d996534b62f7b6b55bfe6eeb9f5fc679af7a
Fixes: 230080345
Test: reproduce topClipAmount glitch while media is present, drag on media

Former-commit-id: aa72d09a5407f1d569628a5e5a0ae14777ab5bb6