Reuse common transition method used by both legacy and shell
To reduce scattered duplicated code of "isAnimating+inTransition"
when the places have the same condition to check either legacy or
shell transition.

The major fix is in WindowToken, which avoids clearing fixed
rotation too early when FixedRotationTransitionListener receives
onAppTransitionFinished for an activity which is still in transition.
E.g. launch landscape app from portrait home, and then swipe up
before the launch animation is finished. The closing animation should
still show the app in landscape.

Bug: 223397413
Bug: 212570341
Test: atest TransitionTests#testAppTransitionWithRotationChange

Change-Id: Ib954b390137ea11ccc45748b62e7b7d6812dceb7

Former-commit-id: 18b51c91fbbae544a79d3d2fadb51cd974b65cce