Rename IInputContext to IRemoteInputConnection
IInputContext is the IPC interface name used behind InputConnection,
which was named in the initial development of IME API for Android 1.5.

To make it more descriptive and consistent with other classes such as
RemoteInputConnection, let's rename it to IRemoteInputConnection.

This CL also renames

  IInputContextInvoker

to

  IRemoteInputConnectionInvoker

for consistency.

This is a rename of internal IPC interface.  There should be no
user/developer observable impact.

Bug: 234882948
Test: presubmit
Change-Id: Iae82d5a26f0746161c2bf91a421b97f0f238aafd

Former-commit-id: e893ec4f8d71f2cbc98ea3368203d92a19d4a0e2