Deprecate IMMS#executeOrSendMessage()
This CL addresses the remaining TODO from Bug 215609403.

With this CL, IInputMethodClientInvoker takes over the responsibility
of what

  InputMethodManagerService#executeOrSendMessage()

has been doing so that InputMethodManagerService does not need to
worry about whether IInputMethodClient is a BinderProxy or not.

This CL is expected to be really safe.  At least for normal cases when
the IME client is not system_server, there is literally no semantic
change.

Bug: 234882948
Test: presubmit
Change-Id: I0a7dd5162c4b94599b896420b8427913fa772414

Former-commit-id: 1112d366b6aaca4ac8996ddfed178e65adfa303a