Remove IInputMethodClient from hiddenapi-unsupported.txt
There has somehow been the following AIDL-generaged method

  IInputMethodSession IInputMethodSession.Stub#asInterface(IBinder)

in hiddenapi-unsupported.  This is strange because A) there is no way
for IME client apps to obtain an IBinder object of IInputMethodSession
and B) even if IME client apps somehow get a raw IBinder object of
IInputMethodSession its IPC methods are not accessible due to
restrictions on non-SDK interfaces.  Therefore, the risk of
accidentally breaking apps by removing this entry is supposed to be
very low.

To make future refactoring possible, let's remove this entry from
hiddenapi-unsupported.txt.

Fix: 235267988
Test: presubmit
Change-Id: If3b0737854038e39c2bfe4f7811c5d7d7bdc18a7

Former-commit-id: 1f0211714c3623c611194dab62d28bd0d54fbed8