Let InputMethodUtils not log NameNotFoundException
This is a follow up CL to my previous CL [1], which aimed to
gracefully handle

  PackageManager.NameNotFoundException

thrown from

  PackageManager#getApplicationInfo().

One thing I missed is that showing warnings in logcat is not that
useful and can be confusing, because seeing the full stacktrace often
makes people think something is terribly broken.  However, in our
case, not seeing packages listed in

  R.array.config_disabledUntilUsedPreinstalledImes

is quite expected and never an error.  Let's not log that.

Anyway, this is just about logging and there should be no user visible
behavior change.

  [1]: I58af03ec96103b60e20758558a92dd9408baab2f
       0a716212488d2aef01d57a98368da3baed888403 [formerly 4654609a7eacf2abe0d8f6cee7126763ecb81ad6]

Fix: 235281743
Test: Manually verified as follows.
 1. Build "coral-userdebug" and flash it
 2. Complete the setup wizard.
 3. adb logcat -c
 4. adb logcat -s InputMethodUtils:*
 4. On another console, run "adb shell ime reset"
 5. Make sure there is no log.
Change-Id: I61bfb558a18383163ddea59c95f66b7246f30ede

Former-commit-id: 6318959d40a0f0d56a32a805e96e7b86ee868fbd