Update InputSink on transition finish
InputSink surface properties depend on whether the activity
is in a transition or not. This means that when the transition
state changes we need to update. Start is currently handled
already via prepareSurface, but finish is not.

Additionally, isInTransition was incorrectly checking if
a transition is ongoing rather than if the activity itself was
in the transition.

Bug: 234636672
Test: atest ActivityRecordInputSinkTests#testOverlappingActivityInNewTask_BlocksTouches
Change-Id: Iff7bc2dfbebe8024a8f666cce9b39bc92fa35a0a

Former-commit-id: 9f1dae8b99e9cf70836aa0249c048b2341de35bb