Fix squished notifications on the lockscreen.
This fixes a stale state caused when stack height is not recalculated after being frozen.  NSSL uses shouldSkipHeightUpdate() to prevent the stack height from being recalculated during various animations and gestures, but we didn't correctly trigger a recalculation after that 'freeze', so this CL adds the necessary recalculation at those transition points to prevent the old stack height from persisting until the next update.

This also fixes a visual jump when unfreezing the stack height if it was frozen while the appear fraction was still changing.  This jump is fixed by allowing the appear fraction to continue to apply and update the stackHeight (from the stackEndHeight) even while the stackEndHeight is frozen.

Fixes: 235366457
Test: drag shade up during aod->ls transition
Change-Id: Ice6de1149b05564294e5c51308e962112cf4eda7

Former-commit-id: 82596c5c5dd6c11da61000a17488615bfe12b01d