Use TRANSIT_WALLPAPER_OPEN when going to home activity
Some devices don't show a wallpaper when the home activity is shown.
This triggers a TRANSIT_TASK_TO_FRONT transition when e.g. pressing
the home button. TRANSIT_TASK_TO_FRONT is normally used when an old
activity is started on top of the old one. Normally the new
activity is put on top of the old one by adjusting the Z order
for the duration of the animation (via mNeedsZBoost).

However, the home activity can not be put higher in z order than other
regular activities. The z boostinga logic defines three groups of
activities - home, regular and always-on-top. The z boosting works only
within a group. Hence, home can't be animated on top of the old
activity, which breaks animations like fade in.

This CL makes the transition to the home activity be classified as a
wallpaper_open transition, which assumes that this is the bottom-most
activity on the screen

Bug: 220024691
Test: manual: start a random activity and press the home button -
      observe animation
Change-Id: Ida53d274ea050cc64d3d27173e32b5b1adb2fe5f

Former-commit-id: 168084eb372aad011b43e2fc29dcc58bd36af117