Move IInputMethodSession to com.android.internal.inputmethod
This CL Moves IInputMethodSession from com.android.internal.view to
com.android.internal.inputmethod to be consistent with other
IME-related files.

This CL also moves IInputSessionCallback to
com.android.internal.inputmethod with renaming it to
IInputMethodSessionCallback for consistency.

This is just a renaming CL of implementation details.  There should be
no user/developer observable behavior change in this CL.

Bug: 234882948
Test: presubmit
Change-Id: I87495cbd482dd55005ac8d948f57f32bb016c7d6

Former-commit-id: fd1580e341470aa355f2edbfdf0dc5e3134c5555