Introduce LocalAndroidColorScheme to expose Android colors
This CL introduces the AndroidColorScheme and its CompositionLocal to
access any Android color from a Composable. The list of colors was taken
from [1].

Ideally we would only use M3 colors but given that M3 and Android
colors have not been aligned yet, our current UI specs are still using
the Android colors.

In the long term, this class will be deleted or slimmed down to contain
only the few colors missing from M3.

[1] http://cs/android-internal/frameworks/base/core/res/res/values/themes_device_defaults.xml;l=214;rcl=fa86dc4ca511d27ae2d32749d180466404c18218

Bug: 230605885
Test: atest SystemUIThemeTest
Change-Id: I45927f23eb8256987b4e5c6ae48e33ce581aabb3

Former-commit-id: 78b3e889ff0a3570fc32838426052f9704d17688