Remove contents from PackageDataSnapshot
This interface is designed to be an empty type provided to non-PMS
callers to pass into package related APIs. It shouldn't have methods
itself to discourage the access of the data manually outside of those
APIs, since the caller may not be aware of how to cache and use snapshots
appropriately.

Bug: 235834042

Test: atest AppsFilterImplTest

Change-Id: I5e48ef006fdcb242bdcb1e6d0272c5b4e5513e11

Former-commit-id: 697894bbd4a7fdd8108a11231c693f552e4e2c3f