[Condition] Subscription Introduction
This changelist introduces subscriptions for condition callbacks.
Subscriptions allow callbacks to specify a set of conditions they are
interested in. In turn, the monitor they are registered to resolves
these conditions into a unique set and updates callbacks accordingly.
This refactor is the first step towards consolidating monitors into
a Singleton.

Bug: 235627154
Test: atest ConditionMonitorTest
Change-Id: I6385b279b965adc1a8713420497c059b891addf1

Former-commit-id: 581e7e03968f6889bb72277d9782f6538104e474