Fix notifications disappearing when shade fully expands.
The notifications would disappear with the shade fully expands because the shade calculation uses the expanded state if it's not 0 or 1, so when expansion becomes 1 the 'fraction' falls back to using the doze amount, which is 1, causing notifications to disappear.  This was only happening because of a bug where the dozeAmount was left as 1.0 while the device was unlocked during the screen off animation.

The core fix here is to clear the dozeAmount when cancelling screen off.

Fixes: 234979264
Test: enable aod & rear/side fingerprint auth; go to launcher; tap power to turn screen off; unlock with finger during the animation; pull shade to see notifications
Change-Id: If781da0cf941d3560be8bd6b27429e1d94448576

Former-commit-id: ef8d78874f344ca344ed4caa369f5a042e7fcc2c