Remove uses of depfile in gensrcs modules
The depfile concept in Soong and Ninja isn't applicable in Bazel because Bazel requires listing the deps explicitly.

This CL removes the uses of depfile property and explictly declare the implicit deps in tool_files. The motivation behind this change is to eventually deprecate depfile (aosp/2120835) in gensrcs to ensure all gensrcs modules are convertable to Bazel.

Test: CI
Test: m framework-cppstream-protos && m framework-javastream-protos
Bug: 179452413
Change-Id: Iec6b61b4194d3f1e728449ee63b879818d9c958a

Former-commit-id: 61e69fabe18856b653a1af253a8206102abed607