Use isVisibleRequested to replace with isAnimating for IME picker dialog
In CL[1] to fix IME picker dialog may still be selected as the focus window
when quick-switch to the next app from the IME layering target.

Since even though the dialog surface is attached to the app where the
surface hierarchy is lower than the next app, the dialog is a system window
where the window hierarchy is higher than the next app.

Previously we use isAnimating to check if the IME layering target is in
transition state, which may costly and not reliable.

In order to check if the IME layering target is leaving more efficiently,
use isVisibleRequested to replace with isAnimating state and it also
benefits when in shell-transition.

[1]: I771d9817cd6dfcb1bf0d576d94bea9d1b9adc80e

Bug: 212570341
Test: atest DisplayContentTests#\
    testImeMenuDialogFocusWhenImeLayeringTargetChanges
Test: manual, ensure Bug 195073688 won't happen with enabling
      shell-transition
Change-Id: I6857c13ab8bbd0ca84d0e145fc783d4e64b8dcee

Former-commit-id: ae6bb5a6d2332eaceb218d7a3eb68284d8656afb