Fix high traffic PMS lockless methods
The notifyPackageUse method is called so many times that it is
causing a spike in snapshot invalidations. This reverts to a
(broken) S state where the code takes the lock and updates
the time value, but doesn't actually invalidate the snapshot,
preventing the re-generation.

Also fixes a check in grantImplicitAccess that was calling
onChanged() even though nothing changed, which was the other
significant method.

Test: manual, run profiling steps in bug on pre-change and post-change
Test: atest AppsFilterImplTest

Bug: 236021119

Change-Id: I20b1bb1d8b4f3df09c901f39db8462ab1a67607d

Former-commit-id: e574a37db187958d57ca2733f40da840ea472a45