Allows system uid to access application across users
After this change, applying the
function shouldFilterApplicationIncludingUninstalled to the APIs
which do not take the user id as the parameter becomes easier.
There’s no need to consider whether the system caller invokes
the API with a correct user id. The system caller is able to
access the application across users.

Bug: 229684723
Test: atest AppEnumerationTests
Test: atest AppEnumerationInternalTests
Test: atest CrossUserPackageVisibilityTests
Change-Id: I674281a1fa9083bcf63dd67cad2f944d7481cb39

Former-commit-id: a4fb7ea1712400c82357a411dccfe00f0b5ea7a3