Reset the keyguardGoingAway flag when keyguard is not dismissed.
LockIconViewController relies on
KeyguardStateController#isKeyguardGoingAway() to determine whether
keyguard is currently showing or not. This flag is set to true when the
keyguard is almost completely dismissed by swiping up, but it is never
reset to false when the swipe up is cancelled at the last minute.

This change resets the flag if the lockscreen is not dismissed.

Fixes: 233845417
Test: manually
Test: atest com.android.keyguard
Test: atest com.android.systemui.statusbar
Change-Id: Ia36b6bb15de87a1b343e62772d52cbc6c760a504

Former-commit-id: 595c164aa6d67d19799fbc37b3ddb972ad5041f6