Remove SharedLog usage from NetworkStackClient
As was pointed out, SharedLog in NetworkStackClient (and
ConnectivityModuleConnector) was only used for dump() functionality
which is not called by anyone. This CL removes the unused dump()
functions and removes SharedLog usage in favor of android.util.Log.

Also cleaning up usage of Slog so only logWtf logs to the system log in
both NetworkStackClient and ConnectivityModuleConnector.

Test: m
Bug: 235901424
Change-Id: Ia484ee825cd714fd8eb7704ca1d16acfbe2e7331
Merged-In: Ia484ee825cd714fd8eb7704ca1d16acfbe2e7331

Former-commit-id: 2d8a3422d2996bbc8b45eef59521e0c710b26f34