Fix a side channel leakage for the pm#getLaunchIntentSenderForPackage
am#isIntentSenderTargetedToPackage returns false if the target intent
data has a non-null package name and a non-null component name. This
condition hits the successful case of the
pm#getLaunchIntentSenderForPackage api, and the application can
detect package's existence information via both apis. Removes the
package name from the target intent if the lauch intent is found
for the package to fix the issue.

Bug: 228450093
Test: atest AppEnumerationInternalTests
Test: atest PackageManagerTest
Change-Id: I5ec66827eae2a2416e0332694affd2f613a5e965

Former-commit-id: 3cae08367a82e34baa6c3fd78276e20b836af282