[DO NOT MERGE] Adding Region Sampler to Animatable Clocks
A RegionSamplingHelper is used to extract the color of the wallpaper
behind an AnimatableClockController. This is used to set the color of
the clock to a contrasting hue so that the text is legible. These changes
fixes the problem of the clock text blending into the wallpaper.

Bug: 202758428
Test: atest AnimatableClockControllerTest
Change-Id: I645c2a0b4e68d5ec9165a2eb354f6d814acbf596
Merged-In: I645c2a0b4e68d5ec9165a2eb354f6d814acbf596

Former-commit-id: 13c1e7eff3b468176400ccfa17be8842f3143566