Deprecate mRequestUpdateCursorAnchorInfoMonitorMode part 1
This is the initial step towards deprecating

 InputMethodManager#mRequestUpdateCursorAnchorInfoMonitorMode,

which was introduced as a minimum implementation to support
CursorAnchorInfo API in EditText [1]

Before this CL, there is a behavior discrepancy in

 MSG_UPDATE_VIRTUAL_DISPLAY_TO_SCREEN_MATRIX

handler because mRequestUpdateCursorAnchorInfoMonitorMode has a valid
value when and only when the editor component is EditText.

With this CL, the same behavior will be applied to non-View based
editors.

 [1]: I3c6b69bd9d79b199afe68d838f25effa6048e5cc
      cf8421df856a5950729ff1d0e73f21d480aa98fb

Bug: 236713697
Test: presubmit
Change-Id: I8d42aef4ffd81253181ba876f8970179643d834c

Former-commit-id: c5af33d324384e427cf3fa97b7574b3cde7d5d90