Deprecate mRequestUpdateCursorAnchorInfoMonitorMode part 2
This is the second step towards deprecating

  InputMethodManager#mRequestUpdateCursorAnchorInfoMonitorMode,

which was introduced as a minimum implementation to support
CursorAnchorInfo API in EditText [1]

Before this CL, there is a behavior discrepancy in

 InputMethodManager#updateCursorAnchorInfo()

handler because mRequestUpdateCursorAnchorInfoMonitorMode has a valid
value when and only when the editor component is EditText.

With this CL, the same behavior will be applied to non-View based
editors.

 [1]: I3c6b69bd9d79b199afe68d838f25effa6048e5cc
      cf8421df856a5950729ff1d0e73f21d480aa98fb

Bug: 236713697
Test: atest CtsInputMethodTestCases:InputMethodServiceTest#testOnUpdateCursorAnchorInfo
Change-Id: Ib17a05e11b7948c9e468e4efd0497dc1ed2f3559

Former-commit-id: 752bf17e6b2c5619dafc96504c1e56900caaf3de