Skip moving the focused Task to top if it is already on top
The Task was moved to front while making an Activity to be the
focused app. In that case, the resumed activity will be reset
to the top-most resumed activity in the Task.

However, the top-most resumed activity in the Task should not
always be the current focused activity when activity embedded.

This CL skips unnecessary task movement and making the target
activity to be focused if the task is already on top and focused.

Bug: 236565088
Test: atest TaskFragmentTest
Change-Id: I6cbea10bdfa422e9d51c051595ed592843f5ba79

Former-commit-id: 9b71e4551cef015b3ccc40e65f0d7ee14d9bb257