Increase the max size of emulated displays
This CL incrases the max display size that can be emulated. The previous
logic would only allow a size that is maximum 2x bigger than the host
device emulating the display. For screenshot diff testing, we need to
emulate big devices (like Lenovo P12 Pro) on small devices/emulators, so
we need to increase this maximum size.

Bug: 230832101
Test: Manual
Change-Id: I4b6b871b75a638a1c818931c5c8970873d886cf3

Former-commit-id: 825dd5d1c68f1bb34540edbf464fa6ac600d0e0d