Fix null handling of getPackagesForOps
This method's "ops" argument is documented as null being an acceptable
value:

    @param ops The set of operations you are interested in, or null if
    you want all of them.

However, passing null in fact always leads to a NullPointerException,
and so the "if you want all of them" functionality is effectively
uninvocable.

This commit allows null to be passed, resulting in behavior that matches
the documentation.

Fix: 182472424
Test: atest RuntimePermissionsAppOpTrackingTest#testGetAllPackagesForAllAppOps
Change-Id: Ieba1f401aba6f5607e89fa8b932c66ddf6b73ad4

Former-commit-id: 374a7d704cd70d5b6c6810b3af85d0c0861536eb