Moving BrightnessInfo and BrightnessReason outside of DisplayPowerController to have the support of declaring the brightness relevant states in DisplayPowerController agnostic fashion.
We primarily need this because currently DisplayPowerController is
tightly coupled with these classes, even though it is not really in the
realm of DisplayPowerController to evaluate it. In future, the vision
is to have a different class which keeps track of all the changes
happening to the brightness and store them in a RingBuffer(Like we do
today). This is taking it one step closer to that, with more verbose
tests
Bug: 233051797
Test: atest BrightnessEventTest && atest BrightnessReasonTest

Change-Id: I15d47cc669bda1c160f02e205168c7aeee9d6658

Former-commit-id: 373794989baaf46d724431694831e9aa388ab783