Progress on permissions sync transport migration.
An earlier change introduced a socketpair() based CDM transport,
which is robust "streaming" interface between two devices.

This change updates the existing permissions sync logic to use this
new transport for its payloads.  It also begins returning success or
failure messages so the requesting side knows when the permissions
have been applied.

We also rename the package and class hosting the transport logic,
and migrate our socketpair() based transport logic there.  Since
the initial version of this protocol won't be using protobufs,
and doesn't need to implement manual pagination, this change also
removes that logic to reduce the size of our backports.

Finally, since this transport isn't ready to be used by third-party
apps, we add security checks to only allow built-in apps.

Bug: 237030169
Test: atest CompanionTests
Change-Id: Ia0dab24d44f1973d5af6ad8a2abb198408a681df

Former-commit-id: 9ac551161b9d6df2f248aa3b016d6c2b665e9da2