Switch NetworkTimeUpdateService to internal API
Switch NetworkTimeUpdateService to using TimeDetectorInternal instead of
the binder API.

The NetworkTimeUpdateService runs in the system server, so it does not
have to use binder to communicate with the time_detector.

Also update NetworkTimeSuggestion to include uncertainty so that it
contains all the information needed for future usecases associated with
GnssLocationProvider.

Bug: 219693030
Bug: 222295093
Test: atest services/tests/servicestests/src/com/android/server/timedetector/
Test: atest services/tests/servicestests/src/com/android/server/timezonedetector/
Test: atest core/tests/coretests/src/android/app/timedetector/
Change-Id: I4665da76429bc333b34769fc362221581472690e

Former-commit-id: 28f3fe363b5f6eac42ebfc740f25ad697c32f4bf