[Split shade][Motion] Fix scrim not being fully opaque after shade expansion
When using fling to expand the shade, the scrim was not at full opacity
at the end of the shade expansion.

This is because we are using a custom expansion fraction for split shade,
instead of reading the fraction provided by PanelViewController.

The fix is to only use the custom fraction during the expansion transition,
and when a completed transition is detected, use the default fraction.

Test: Manually
Test: ShadeTransitionControllerTest.kt
Test: ScrimShadeTransitionControllerTest.kt
Fixes: 237101056
Change-Id: Ifa87309347d501aeb293d6e2f01d537943996b4b

Former-commit-id: 39d63e51973298b8dfd5ccbe52f31bbac099134b