Replace Somnambulator functionality with shell commands.
This changelist removes the ability for Somnambulator to enter dreams.
Somnambulator itself was removed from use in ag/238189, where this
functionality was moved to UiModeManager. Somnambulator was maintained
to resolve the dock intent (but not necessarily launch) and also to
be a way for launching dreams for development/test. This changelist
preserves the intent resolution, while providing the ability to launch
dreams through a shell command as root.

Fixes: 225878553
Test: verify dreams work through docking with change.
Test: adb shell cmd dreams start-dreaming
Test: adb shell cmd dreams stop-dreaming
Change-Id: Ia88de2e1548c12079690860a2bb75e4e79d10e66

Former-commit-id: 74c52366701debed5f39c629ab85906fd3965605