Use layer mirroring when screen recording via VD
When a Display has the flag VIRTUAL_DISPLAY_FLAG_AUTO_MIRROR, the
content on the display is expected to mirror another display. This was
previously done by setting the same layerStack. This change removes that
behavior and instead will create a mirror hierarchy of the other display
and place it at the root of the new Display. This is leveraging the work
done for ContentRecorder, but using it as default for all displays
instead of only for VD created by MediaProjection.

Test: Screen record via MediaProjection
Test: Screen record via DisplayManager
Test: ContentRecorderTests
Test: DisplayContentTests
Bug: 237664947
Change-Id: I18b36dadc43ba1c37f9516ff10d24bb8683c04ac

Former-commit-id: e64bae455ed9230b68afa6fbbfd825b8d2988894