Fix the visibility check for sdk_sandbox processes
For some cases (e.g. WebView), the visibility is implicitly granted to
the calling uid. Right now the visibility rules for sdk sandbox uids
only take into account the forceQueryable apks, which means that WebView
will never be visible to sdk sandbox uids.

This change adds the implicitly queryable apps to the app visibility
check for the sdk sandbox uids.

Test: run SdkSandboxClientWebView
Test: atest CtsSdkSandboxInprocessTests
Test: atest AppsFilterImplTest
Bug: 238871061
Change-Id: Icfeb8e48151e1675d7b85c9714238f4a499ec6ea

Former-commit-id: 5a1cba22d1e6a4cc9fd549ef1f9216f38bfdc601