Fix notification animation from AOD
This CL fixes the notification animation when launched from the AOD by
making sure that we don't play both the notification animation and the
wake-and-unlock transition.

As shown in b/236487811#comment24, the resulting animation looks better
but still not perfect. A better, longer term fix would probably be to
not wake up the phone at all, do the transition, then instantly hide the
shade/keyguard.

Bug: 236487811
Test: Manual
Change-Id: I4dec3651ea385a74c3564ad7c19fb10802b3464d

Former-commit-id: b465763e12c7edc59ac720b50d7761cfa2ed8e29