Move most of the shade codebase into its own package
Creates the package outside the 'phone' directory, since the shade is
not just on phones anymore and moves most of the shade code to it.
Moves the shade.transition package to it. Moves all related tests.
Expands a small amount of method scoping as necessary. Updates, adds,
and organizes imports as needed

Test: atest SystemUITests SystemUIGoogleTests
Fixes: 238333941
Change-Id: I721d35d4adb637f06501e92efe2d37421a3e84e5

Former-commit-id: 8a75e4922ac079e8ca293a78de43a2aef941948d