Clean up the LockSettingsStorage file methods
For file paths, consistently use File instead of String, and don't use
ambiguous variable names like 'name' and 'key'.  Remove unnecessary
conversions to absolute paths, as the paths are all absolute already.

Consistently use 'data' when referring to a file's data, instead of
'hash', 'stored', or 'value'.

Don't catch exceptions from deleteFile(), as it doesn't throw
exceptions.

Remove the DEBUG field, as it's hardcoded to false, and it only enabled
the printing of two messages which were redundant with each other.

Test: atest com.android.server.locksettings
Change-Id: If0dd16d8c16015d129bda75e93600e60a1341f82

Former-commit-id: 2515cd5c161e7fd3686b16c9e04596892272cb1f