[Keyguard] Fix unlock animation for occluded LS.
When we open shade from dream (which is in an occluded keyguard state)
and open settings. We are prompted to authenticate with the bouncer.
Upon doing so, the app hangs in keyguard even though the user is no
unlocked. In order to fix this, we opt not to defer keyguard done when
keyguard is occluded.

Test: Manual and a unit test
Bug: 228064734
Change-Id: I58aebe033f069dc6bedb3ad3a78da849ed29dc78

Former-commit-id: 6096b1a59705fd3cdec617c9a95f3cd00935ab76