Add error log for when widget provider is null
When the service try to update the visibility of the widgets, it would
iterate over the available widgets and check the ids of their providers.
However, due to unknown reasons there is a possibility of their
providers being null even after they are bound to the screen. Adding
more logs to help determine the exact issue.

Bug: 212651639
Test: N/A
Change-Id: I74a095dd2a44d0235fba27a57190aa89032fa63e

Former-commit-id: 5363de2f63e158eaa853b29b1d1c009b99b01514