Undo parent's scale for child window with different override scale
An app of a different package can add a child window to another app's
window. If the app of parent window has override scale, it will draw
its content in a scaled environment. But if the child doesn't have
override scale, it still draw in with natural size of the device.
Then the child surface should still show without scale, so the scale
from parent surface should be reversed.

Bug: 182362657
Test: atest WindowStateTests#testCompatOverrideScale
Change-Id: I83b32090ceb2e92d644e83b9b14ea68ef2ca96f2

Former-commit-id: 6565924b5e484acf9aae0814e20f4284896979ba