Add support for device emulation in the Gallery app
This CL adds support for device emulation in the Gallery app so that we
can easily emulate any device on another device, e.g. emulate a phone on
a tablet or a emulate tablet on a phone.

The emulation is done at the device level, so that might even be useful
outside of the Gallery app, e.g. when working on SystemUI.

See b/231131244#comment11 for a quick video.

Bug: 231131244
Test: Manual
Change-Id: I64094369c32b796ad41a274d23e0910b7c61a3b0

Former-commit-id: a38f25ded64f61973d745c67fcdd107501c1d42f