Prevent displays that are disabled from receiving input
For foldable devices with two displays, both displays often map to the
same layer stack. In this case, the we expect only one of the two
displays to be used at once, and the other should be disabled. When the
device is folded/unfoled, the display that is enabled will change.

When two displays map to the same layer stack, it is currently difficult
to identify the display which the layer stack should be associated with.
Picking the incorrect display will result in issues with input
dispatching. For now, we get around the one-to-many mapping between
layer stacks and displays for input by preventing displays that are
disabled from receiving input, assuming only one display is enabled at a
time for a given layer stack.

Bug: 238122363
Test: None
Change-Id: I7cc5c4296bb8d891c89fc334f13988dd39d936f3
(cherry picked from commit 2b3251cb61436dcae97659f7f656a4bd8e457fa2 [formerly 3da19deb307571959324fd27ef0a9215e314cb6b])

Former-commit-id: 455ebf5e04e68c767407128fef31fdffa62638a3