Make IMMS#canShowInputMethodPickerLocked() self-contained
With this CL,

  InputMethodManagerService#canShowInputMethodPickerLocked()

becomes the only and unified verification mechanism about who can call

  IInputMethodManager#showInputMethodPickerFromClient().

This CL also fixes an edge case when a background IME process is
allowed to call that API when all the following conditions are met.

  A. The calling process has the same package name as the current IME
     for the foreground user, and
  B. The calling process has INTERACT_ACROSS_USERS_FULL permission.

It is quite unlikely that existing IMEs have relied on the above edge
case, because INTERACT_ACROSS_USERS_FULL is a system API that should
not be granted to IMEs in general.

Either way, with this CL, future readers should be able to understand
how the caller is verified more easily.

Bug: 34886274
Test: presubmit
Change-Id: I423b109bd55c13397e9236890ce72d043d1b7a89

Former-commit-id: 1d0387f0282adb9c91ce480cb3fe9fb883aae3c6