Remove TODO item in IMMS#canShowInputMethodPickerLocked
As CL[1] has already supported to show IME picker on the
external/Virtual display with the self reported displayId of the
current focused client, so currently canShowInputMethodPickerLocked
for Single-Session IME usage with switching focus displays
should be enough to check if the given client is the focused client.

Remove the TODO item to prevent the misleading of not support
multi-display for IME picker.

[1]: Ic7d7c5a7ad8005a3fbd9d1c1b73e3c5a39a07001

Bug: 238972096
Test: atest InputMethodPickerTest
Change-Id: I2bdf3c2a8325d95025196782b9fdaae7441124b7

Former-commit-id: de5aa37a17df7d1068f0877bf26c69ad5ddaa2e4