Update system server profile based on input method cleanups b/234882948
Some package names and class names have been changed. This CL manually
updates the profile to make it in sync with the changes.

The motivation is to address the performance regression caused by
profile-guilded compilation. This manual update is on a best effort
basis. There are many methods in the input method code that have been
added/removed/renamed, and this CL does not reflect those changes.
Therefore, the performance cannot be exactly the same as before unless
a new profile is generated from dogfooder's data.

Bug: 238413831
Test: -
  1. m
  2. out/host/linux-x86/bin/profman \
       --create-profile-from=frameworks/base/services/art-profile \
       --output-profile-type=app \
       --apk=out/soong/.intermediates/frameworks/base/services/services/android_common/aligned/services.jar \
       --dex-location=/system/framework/services.jar \
       --reference-profile-file=out/soong/.intermediates/frameworks/base/services/services/android_common/dexpreopt/profile.prof
  3. See much less "Could not find method" on
     "Lcom/android/server/inputmethod/InputMethodManagerService;"
     than before.
  4. out/host/linux-x86/bin/oatdump \
       --oat-file=$ANDROID_PRODUCT_OUT/system/framework/oat/x86_64/services.odex \
       --dex-file=$ANDROID_PRODUCT_OUT/system/framework/services.jar \
       | grep "hideSoftInput" -A 100
  5. See native code.
Change-Id: I52144b00b752c78be476be07e2ecd453e1e252cf

Former-commit-id: c2b7c52221871c8da2936fb3aa34174b14050b24