Ignore the Recents animation target when collecting keep clear areas.
In phone it is possible that user e.g. swipes up to close full screen app / go home, and there could be underlying windows beneath it that are trying to register keep clear areas, but due to window ordering and top-to-bottom lookup, they would never get a chance to report them and trigger the onKeepClearAreasChanged, since the full screen app is still on top of the list and the loop exits early. However, that app is simply being animated and should no longer impact this logic.

Test: atest DisplayContentTests#testKeepClearAreasMultipleWindows
Test: atest WindowStateTests#testKeepClearAreas
Bug: 183746978

Change-Id: I735a9cfbc9b069d22709e96c61f776104fea1501

Former-commit-id: 04c355ad8efb0845fda7fa399c5b19f13fe859a5