Track interfaces and roaming changed event history
Currently netpolicy event log includes the meteredness changed
event but it's also called while the roaming changed. The
roaming event should be tracked as well so add roaming event
for it. Besides, it would be better to track the interfaces changed
so this change also adds interfaces event for tracking interfaces
changed history.

Bug: 237640200
Test: FrameworksNetTest
      adb shell dumpsys netpolicy
Change-Id: I788fb8896d3cd5b47916c687fde05ef7ad88456b

Former-commit-id: a58a05a403aa357f26122c56edcd070f803887aa