Bouncer should dismiss if being tracked when face unlocked
When the user is tracking the bouncer with their finger and a
successful face unlock occurs, the bouncer should always be
immediately dismissed. Prior to this, it was possible for the user to
continue tracking the bouncer over the homescreen after unlock, leading 
to shade problems, including missing lock/udfps icons and
lockscreen over homescreen/activity.

The primary issue was BiometricUnlockController was returning MODE_NONE, and 
is now returning MODE_DISMISS_BOUNCER.

Fixes: 237822716
Test: atest StatusBarKeyguardViewManagerTest
Change-Id: I364d8e61a61138d79240b5baefed295493678063

Former-commit-id: 80771ec5636dd5a4f95e06b1de31549341fb3ae3