Add Region Sampler to Lockscreen Smartspace
This change adds region sampling to the views for lockscreen
smartspace. It also modifies how the region sampling instance
is created. The clock parameter is replaced with an interface
that specifies how the updating of colors should be done. This
allows for a more generalized usage of the region sampling
instance class.

Bug: 202758428
Test: atest LockscreenSmartspaceControllerTest
Change-Id: Id890309c95eab1251d09485ce560e1eae5c15900

Former-commit-id: a57e4ce3ed33e9e9b7bbf99dca2029ec8ea33aa5