Change the allowed lower bound for date changes
Change the allowed lower bound for date changes to 2015. The previous
lower bound of 2007/2008 is a long time ago. This is most likely to
affect automated / manual tests, not real user needs: real users should
have no need to set the clock that far back, and may cause TLS issues if
they do.

This is primarily a proof of concept and can be rolled back easily if
needed. No CTS tests for these bounds are planned.

Bug: 228967927
Test: build only / manual inspection of SettingsUI
Change-Id: I9b8ba632038c7d02d630f96da27303f5d0d109fc

Former-commit-id: ae911cc91d47acdc4120cbafcc5d914383114767