Fix test dependencies for AppEnumerationInternalTests
All of a test's dependencies should be explicitly declared in the Soong build module to enable infrastructure features that ensure correctness and optimize runtime. Here in the test config file, AppEnumerationInternalTests uses `AppEnumerationSyncProviderTestApp.apk`, so that this change declares `AppEnumerationSyncProviderTestApp` in the Soong build module.

Besides, this change modifies the test config file by switching to use `PushFilePreparer` instead of `FilePusher`, since `FilePusher` is designed for the tests that are part of compatibility suite and AppEnumerationInternalTests is not.

Bug: 222438068
Test: atest-src AppEnumerationInternalTests \
          --experimental-device-driven-test \
          --experimental-java-runtime-dependencies \
          --bazel-arg=--test_timeout=48000
Test: atest AppEnumerationInternalTests
Change-Id: I8f09efe82b8dbd11d900752c3810d9215026062e

Former-commit-id: 55f2c7f1a13cc933f4047f0736b536a04918480e