Support display resize animation for shell transition
Apply scale to screenshot layer according to the display size.
Use sync transaction for display traversal, so when the logical
display is changed, the screenshot can cover the change at the
same time. The windows (e.g. display cutout) above the screenshot
layer are also requested to use sync transaction, so before the
sync start transaction is applied, the intermediate buffer won't
show on screen.

Bug: 223397416
Test: adb shell setprop persist.wm.debug.shell_transit 1; reboot
      Enter Settings -> Display -> Screen resolution
      Observe no flickering when switching resolution.

Change-Id: I30b147a220fa2e0b18f0948baa4e65436d995623

Former-commit-id: 7d03d85d8e5998824ea83e54fbe9e7bf71a36d0f