Add REQUESTED_PERMISSION_IMPLICIT flag
This adds a new permission flag, REQUESTED_PERMISSION_IMPLICIT, which
denotes that a permission is an implicitly-requested permission.

The concept of an implicit permission already existed (we just didn't
have a flag for it) and means that the permission acts as a "requested
permission" despite not being explicitly requested in the
AndroidManifest.xml as a <uses-permission>. For example, split
permissions result in implicit permissions, as well as a few unique
permissions.

Bug: 195016052
Test: atest ImplicitFlagTest
Change-Id: I5aaff487d653767b50fda7860d6fd0434a088dee

Former-commit-id: d049e2d73a2095b2cc1f5f3d533a546d1641eb25