Ensure PIN text view stays on screen regardless of font size.
This was previously only anchored to the top of the "1" button, so
ConstraintLayout didn't care if it fell off-screen. This also adds a
constraint to pin it to the top of the parent container (with a bias to
drop it to the bottom of the available space); this ensures that it
remains on-screen.

With font/display density set to max, the buttons now get a little
smaller, but the text view is still on screen.

Test: Manually verified.
Bug: 236950644
Change-Id: Iac0209e297c6c80af616199637871a427184a1ad

Former-commit-id: c193051908561b95033132c50cea7df5bf954b30