Request rollback resistance for SP protector keys
When supported, use rollback resistance for synthetic password (SP)
protector keys, i.e. the Keystore keys that encrypt the SP for a user.

This allows SP protectors to be securely deleted on devices that don't
support Weaver, but do support rollback-resistant Keystore keys.  Secure
deletion of SP protectors is necessary to guarantee that user data
cannot be accessed using old LSKFs or deactivated escrow tokens.

This also aligns LockSettingsService with vold, which already uses
rollback-resistant Keystore keys (when supported) to encrypt all storage
keys that aren't bound to the SP.

Bug: 239632930
Test: atest com.android.server.locksettings (on Cuttlefish)
Test: On Cuttlefish, which doesn't support rollback-resistant keys, set
      an LSKF and checked for the expected log message from
      SyntheticPasswordCrypto ("Rollback-resistant keys unavailable..").
Test: On Raven, which supports rollback-resistant keys, set an LSKF and
      checked for the expected log message from SyntheticPasswordCrypto
      ("Using rollback-resistant key").  Also made some temporary
      changes to log the rollback resistance property of the key, and
      verified that it really got set.
Change-Id: I05f3b7e5c139471febe5c266a39e3dc3bca4831f

Former-commit-id: 0f7dea9083a8ce11bf876ce4f8610bb286af6a2e