Remove redundant enforceCallingPermission()
This CL aims to simplify security checks for

  IInputMethodManager#onShellCommand().

Basically the calling UID check [1] remains to be the main security
policy for that IPC method.  Context#enforceCallingPermission() checks
added in later CLs [2][3] are just redundant with an assumption that
the shell package has those permissions.

For simplicity and readability, let's keep only the most relevant
security check for IMMS#onShellCommand() method.

 [1]: If87189563ccaacd4f9c666bab4f9ad08a9343084
      e9e28d7210c2ff458f62c4dd40ce638b312b2d02 [formerly b8d240fa3f96b7b4ea35dd271beda789044d63ab]
 [2]: I6fd47b5cc1e7da7222774df20247a2c69a70f45b
      bb68f23e7a07c0694ca6ab6234f338f270668e98 [formerly db25df71be4c1bcc654f69ce9a8bff7e3ef46360]
 [3]: If91137e0c144d90ac3046f30607e36bb957a7e0e
      354de6294b17ae15d60afc0bb2894fbff7805138 [formerly 8a2b96bb0a1d901beb9cc9635c1fe9a265799b36]

Bug: 34886274
Test: atest CtsInputMethodServiceHostTestCases:ShellCommandFromAppTest
Change-Id: I19a6833bb7ff6100d4e9428243005377a7fbf432

Former-commit-id: ecac6fd64e9ca4cae452e33f6756543422f42762