[Partial Screensharing] Add a temporary entrypoint in SysUI screen recorder
Adds a button that allows to start app selection flow in
SystemUI screen recorder dialog (see screenshot in the bug).

This button is needed for teamfood testing of partial
screensharing and will be shown only when the feature flag
is enabled.

It will be removed when redesigned dialog with a dropdown
will be implemented.

Bug: 239422418
Test: turn on feature flag and start partial screen recording
Test: atest com.android.systemui.screenrecord.ScreenRecordDialogTest
Change-Id: I9caa913ef52eeb2203dca32e94e0b10d71fc4b83

Former-commit-id: f9ea202f424646f75543afbb1b3c9f7dda262150