Force the app dark/light theme in the Gallery app (1/2)
Before this CL, the Gallery app would use a boolean isDarkTheme that
would be toggled and passed to the SystemUITheme to force set the
dark/light theme of the app. However, this does not work for Views and
the AndroidColors, for which we need to set the app dark/light theme
using UiModeManager.

Bug: 238993727
Test: Builds
Change-Id: Ic3efb42d371c6aac8c2a6297a0cb192a780903cd

Former-commit-id: 33470b902e25b7e4ae6c4ade4c2120b49bd5364d