Combine AppOpsManager arrays into one
AppOpsManager currently has multiple arrays which are all related:

- sOpToSwitch
- sOpToString
- sOpNames
- sOpPerms
- sOpRestrictions
- sOpAllowSystemRestrictionBypass
- sOpDefaultMode
- sOpDisableReset
- sOpRestrictRead

Despite being different arrays, they're essentially a single list of
objects with each field sliced into a different array.

As the entries in these arrays grows, they are are difficult to keep
track of. To simplify things, we can simply join them all into a single
list of objects. Because there are so many fields, we use the builder
pattern to prevent the array from becoming unreadable.

Test: presubmit
Bug: 242097113
No-Typo-Check: typo EMERGECY is preexisting and non-trivial to fix
Change-Id: I1c54a18fc8ca2de6c5ab4a2ae62d1943fcbaea62

Former-commit-id: 5f2bc22d36a48380b1a6585472b7457b43e2f3b5