Fix translation session id collision bug.
Currently the session id is genrated from a static random number
generator created with new Random(), all has the same seed value
in the given Context, so the sequence it generates is identical
across processes

Ideally the id should be generated from a center controller to
avoid collision, the content capture met the same problem before.
Follow the same solution used in content capture, we use SecureRandom
which produces non-deterministic output. But it still good to come out
a more stable solution in the future release.

Bug: 239385816
Test: manual
Change-Id: Ie75acfa6a55a58e0c0a709f2a5ae1c6e5d6dc339

Former-commit-id: 87ff6d9dd0fd87515e43c7e979f4f5b3b89a0c97