wifi: Fix unnecessary notification after restoring config from cloud.
The persistent randomzied MAC address is generated by hash function with
key Ssid. It might be different in different device even if the key
doesn't be changed. But the persist randomized MAC  is not a user setting,
it is generated by the frameworks. Ahe user would need to set up the ACL
again anyways even if they use factory MAC. So it should not to notify
user configuration is changed since it is not a user setting and user
aware it since user is using a new device.

BUg: 235591501
Test: Manual test, no notification show up.
Change-Id: I08fef497df0c42115ee61e92a22009d32666da41

Former-commit-id: 9133f77f1f0d56eda344fdc62ab11a97293265c1