Stabilize quickswitch tests
Previously the tests relied only on TAPL to know when the quick switch was over. However, this is not reliable since TAPL can't differentiate between the snapshot starting window and the app window, causing flickers. This Cl introduces the wait for the app window to be visible (and snapshot or splash screen not to be visible) to stabilize the test

Bug: 240264838
Test: atest FlickerTests
Change-Id: I44772a47a9fd0078ea4618f9c1523c2a23e07f14

Former-commit-id: 1c5e0589d75b97caaa25dfbef91c1e3637b09168