Inline IMMS#startInputOrWindowGainedFocusInternal
This is a follow up CL to my previous CL [1], which aimed to be a
logical revert of changes mentioned in the commit message.

One thing that CL forgot to handle was

  InputMethodManagerService#startInputOrWindowGainedFocusInternal,

which was mentioned in one of our previous CLs [2].

This CL inlines that method again for simplicity.

This is a mechanical refactoring.  There should be no behavior change.

 [1]: If16ac0de536d9089eb04f6e07b1ee47378124658
      d18caa099e1ed167f1c99f31e332babb879d8ca4 [formerly 662b48b72d841119160ab5c1b03be1ea36f3351d]
 [2]: Ic584203c1221fbae17f5e2d8f09e3992df061646
      5be1008b0797b7763cdba508461c96c2a89b80b1 [formerly 5e2d9f271d1f24fc946661579628fb9a817143dc]

Bug: 234882948
Test: presubmit
Change-Id: I7c775245882815483f29acfd83f53b08a40d07ed

Former-commit-id: 324b546ab6288302f5d187fcc26b968d8b4fd74a