If the update is non-interruptive don't re-bubble it
I haven't been able to reproduce this outside of tests on acloud
devices, but if you visit a bubble for the first time (causes an
update to the notif) and then dismiss it sometimes the user dismiss
can happen before  we get the update back from notification manager
service and the bubble will get re-bubbled incorrectly.

To fix this, if a bubble is in the overflow (or active) and there's
an update to the notification that is non-interruptive, we shouldn't
re-bubble that content because there wasn't a new message, just some
change on the notification.

Bug: 238896626
Test: atest BubblesTest
Test: atest --no-bazel-mode --iterations 20 PlatformScenarioTests:android.platform.test.scenario.sysui.bubble.ExpandAndDragToDismissTest
      (with vadim's changes & other CLc)
Change-Id: If0b7b3d3d1875ddd1df6a1412db8c42456de35c5

Former-commit-id: 5c5b6218e41cf3fafae745715a931ca9f4043707