Remove IMMS#calledFromValidUserLocked() as it's unused
This CL finally removes

  InputMethodManagerService#calledFromValidUserLocked(),

as it is no longer used.

This method was originally introduced for a minimum multi-user support
for IMMS [1][2].  The motivation was to allow SystemUI (this includes
the lock screen) and some special processes that always run user 0 to
call at least the following IPCs defined in IInputMethodManager.

 * addClient()
 * removeClient()
 * startInput()
 * windowGainedFocus()
 * showSoftInput()
 * hideSoftInput()
 * showInputMethodPickerFromClient()

As a result, we added the following allow-rules as a quick workaround.

 * callingUid == Process.SYSTEM_UID
 * INTERACT_ACROSS_USERS_FULL

The problem was that IMMS#calledFromValidUserLocked() had been called
not from those IPC entry points but also from other IPC entry points,
and it became really difficult to understand what IPCs must have had
such allow-rules, and what IPCs must have not.  This is basically why
we started an effort to deprecate IMMS#calledFromValidUserLocked for
better code readability.

Anyway we have finally finished the audit and migration, and
IMMS#calledFromValidUserLocked() is now ready to be removed.

This CL itself has no behavior change, since the method to be removed
is no longer used.

 [1]: Ib23849d352db33f0747aa9d5a178f00ac726c13b
      05e8dddd2ffa127a9cb47ebc998797ead4bb6c6d [formerly 4e1ab15b305aac26ad8819fc3b2951e20985944d]
 [2]: I1620413578b9e8da6564664219f65bdc00d5ecfd
      9b6533c0b54e4dcc7b235bcca27bf1ec3bb3daa2 [formerly 135e5fb71242b1151929e2ea7bf221ff421e6ad2]

Fix: 34886274
Test: presubmit
Change-Id: I71a310eea393c8705dc0714a9a968647d76c81ac

Former-commit-id: c354b1948a9311a007b213e30dc9e6ef4c5f46bf