Allow system info to be set visible even if we don't animate
CollapsedStatusBarFragment uses disable flags to show and hide the
system info in two cases: when showing the keyguard and when showing
quick settings. This animation causes the alpha to animate to zero AND
the visibility to be View.INVISIBLE.

This means that the system event animation only has the context to set
the alpha of the system status area -- _not_ the visibility.

In the case where the disable flags change while a system status event
animation is in progress, CollapsedStatusBarFragment would bail on the
_entire_ animation, including the change from View.INVISIBLE to
View.VISIBLE, which is crucial for unhiding the view eventually.

This change simply allows CollapsedStatusBarFragment to set the system
info area back to View.VISIBLE even if it doens't control the overall
animation.

Test: atest CollapsedStatusBarFragmentTest
Fixes: 238968127
Fixes: 239508705
Change-Id: I7ad8c4a4df8ff335ea9b3de428521b24e59002ee
Merged-In: I7ad8c4a4df8ff335ea9b3de428521b24e59002ee

Former-commit-id: 3b218416e00499480c6789252ae71783721b93d9