Clean up protector key terminology
Introduce the term "protector key" to refer to the Keystore key that
encrypts the spblob file for a protector.  Consistently use it
everywhere.  Previously these keys were variously refered to as
synthetic password keys, blob keys, or simply Keystore keys.  This
aligns with updates I've made to http://go/android-locksettings-design.

Reduce the visibility of PROFILE_KEY_NAME_* and
SYNTHETIC_PASSWORD_KEY_PREFIX (the latter is also renamed to
PROTECTOR_KEY_ALIAS_PREFIX), as they are no longer used elsewhere.

This is a follow-up to change I564a738119a47a31b4822d26c6405249f8ce1c06
that addresses one of the code review comments.  No change in behavior
intended; this is just a cleanup.

Change-Id: I69184f7b47b17f0ab25cfa939bf92feb727ba220

Former-commit-id: 995c75399eb63032ec79008746028dd3c974ce14