Actually trigger back action on Back Arrow fling
Currently, when the New Back Arrow is flung and the finger lifted, we do
not actually send the trigger event -- but we do if the finger is lifted
after coming to a "stop", i.e., a commit without a "fling".

To test, swipe from the right edge inwards, and release while your finger is in motion (as opposed to "stop moving, and then lift").

Bug: 240272871
Test: Manual
Change-Id: I81d6e6b1bb42c2cd12d14ae861ba8eb9f3081d11

Former-commit-id: 15ebdf1e7c9722cac448e500d180f11d7d6b069b