Rename pwdToken to stretchedLskf
stretchedLskf is a much more self-explanatory name for what this value
is: it's the result of passing the LSKF through a key stretching
function (scrypt).  The words "password" and "token" are already
overloaded in this code, so this should reduce confusion.  As usual I'll
be updating http://go/android-locksettings-design accordingly.

Test: atest com.android.server.locksettings
Change-Id: Id57de2d1f9246b3aed09554841ecfb54c947582f

Former-commit-id: 174d8dffa9c5cde1983d0ea9bf6d728d6e0467f0