Do not split the suspended packages intent into multiple ones
Previously, we splitted the intent into multiple ones to prevent
from the disclosure of information of packages. This approach impacts
system health. A lots of intents were sent when hundreds of packages
information stored in the intent extras.

This cl stopped splitting of intent. Instead, it passes a filter to the
am#broadcastIntent api for intent broadcast to filter the intent extras
by using the rules of package visibility when passing the intent to the
receiver.

Bug: 238448280
Test: atest AppEnumerationTests
Test: atest SuspendPackagesTest
Test: atest SuspendPackageHelperTest
Change-Id: I6c4051168b855155b49b0bd6b19e0de7dc826df7

Former-commit-id: 04cded96f50da20ef3042492688029e60c68f44f