Turn off the new back arrow in Teamfood.
We're still in the middle of making fairly non-trivial changes to the
new Back Arrow, and it's expected that there might be regressions, or
misc changes that affect its current look+feel.

To avoid confusing folks with such random changes, we're disabling it in
Teamfood for now. The medium-to-long-term goal is to migrate the
UX prototype from its plugin incarnation into production-level code that
lives in this repo (which is when we'll re-enable it).

Bug: 241248730
Test: Manual
Change-Id: I1918008f15f47cfb0f3378462bc168d0bcd8e485

Former-commit-id: d030be309aa1ada00b6598b401c65a7829e6405d