CameraService: Check if the camera is disabled via device policy by user.
The current approach to checking the camera device policy is multi-user
agnostic. This patch takes the user into account.

Modifications to the DevicePolicyManager are required for this to work.
The camera service user needs to be able to make cross-user queries, so
a check for its uid was added specifically for getCameraDisabled.

Test: Tested GCA with device policy set via TestDPC
Bug: 230026863
Change-Id: I922def3310f39b41dade06d656569222f4db46d7
Merged-In: I922def3310f39b41dade06d656569222f4db46d7

Former-commit-id: 92de002983a8c8c3db42b139dc5b9c1a4684938c