Use same height for start and end of date view
This prevents a bad measure (possibly due to optimizer in version of
ConstraintLayout) happening sometimes right before collapsing QS.

Without this patch, right before reaching the end of the collapse into QQS the date would be measured with the layout height for progress 1 (instead of 0). This would cause it to be shorter than it expected to be for the container and therefore not aligned properly (it would badly calculate its top spacing).

Bug: 215584502
Test: manual, fling collapse QS
Test: atest com.android.systemui.shade

Change-Id: I6c9a4425f15408e63adc82d31bad3ae46beebd0f

Former-commit-id: 6c135b7b8cfe3b584cdbf9a5043181bb8b683933