Only send uri in share intent if present in clip
Currently we always set the EXTRA_URI value in the share intent,
even if a uri was not present (null) in the original clip. This
causes some receiving apps to throw an error, since they're
expecting content from the extra that isn't actually present.
Update to only set the extra if we're actually sending content.

Bug: 239854368
Fix: 239854368
Test: sent text content to Keep, verified that "Error importing data"
toast appeared (before the fix) and did not appear after

Change-Id: I39813d9eb512209bc0164c1ecc2fd09e823dc53a

Former-commit-id: 4039ba86ff3f9eb52f330608c3ffcb972d439c57