Provide only caption and IME insets to freeform windows
System bars shouldn't provide insets to freeform windows. Currently,
gesture insets and tappable elements insets by system bars are provided
to freeform windows. These insets are attached to display, and intended
to be used for layouting fullscreen apps.

With this change, only caption insets and ime insets are given to
freeform windows.

Note that gesture and tappable insets from a caption bar are added in
InsetsState#processSource. Thus, this change doesn't affect it.

Bug: 234144193
Test: android.server.wm.WindowInsetsPolicyTest in freeform device
Change-Id: Ife41a37082a65db50c7bb5b75690db5bc3cb0bd2

Former-commit-id: 654386dd0c0e2825f9f925db509053044c1f7318