Moves quick affordances into domain layer.
After much discussion, it became clear that it would be better to move
quic affordance config definitions out of the data layer and into the
domain layer.

This CL does the following:
1. Moves config definitions to the domain layer
2. Moves the "configs" class to the domain layer and renames it to
   "registry"
3. Eliminates the quick affordance repository from the data layer,
   merging its logic into the observe quick affordance use-case
4. Updates the documentation

Fix: 241681067
Test: Unit tests updated. Manually verified the bottom area quick
affordances still work properly when locked, unlocked, and dozing.
Played with the settings under Display > Lock screen to turn the on and
off.

Change-Id: I4b293eeef02e5546fc834008b096e4a7935657fe

Former-commit-id: 6b4f8b1adfd39c344d0d58e259a3d028c51319f7