Fix reactivated media card when recommendation is invalid
Recently active media should show whenever we get an active
recommendation card, even if it is invalid. This was not happening and the
media card would only be visible if the recommendation card was valid.

This adds a condition to hasActiveMediaOrRecommendation that checks
whether a card was reactivated, so that MediaHosts calling it get the
correct visibility.

Fixes: 241150583
Test: manual (steps in bug)
Test: atest MediaDataFilterTest
Change-Id: I5347367bc86a9ca499a47b3ef310e9e5774d79c5

Former-commit-id: 7173943e6380c58c707c5dbf7fca674ab7d62f12