Fix display size calculations in large screen (2/3)
The fix is taken from AOSP patch aosp/2154236.

This will make sure the UI Automator always has the entire display to
interact with. Otherwise, in multi-windowing mode or on devices with
task bar, it will only be able to interact with the current focused
window, the default launch bounds, or the display excluding task bar,
depends on the actual scenario.

Test: PlatformSenarioTests
Bug: 238981445
Bug: 238985243
Bug: 238581838
Bug: 233945217
Change-Id: I9223ef5552ddb081e75de1242670862f2813e56d

Former-commit-id: d1e350c7fce4dd16ec17cbbf717a0fe225af6a7d