Detect layout change of paramsForRotation precisely
WindowManager.LayoutParams doesn't override equals(). It returns
false if the instances are different. So after writing/reading
from parcel, even if the fields are identical, the comparison
of Arrays.equals(paramsForRotation, o.paramsForRotation) always
returns false.

This is the first step as a signal to decide whether the screen
configuration needs to be recomputed according to if there is a
layout change from decor windows.

Bug: 238399969
Test: atest FrameworksCoreTests:android.view.WindowParamsTest

Change-Id: I82582ff48064bbf009403d2c0980a69157860f90

Former-commit-id: e59483dc1c57ed6019edefafb7410da90c04fe87