Logging second order reasons for process starts
Currently the reason is described as the hosting type and name
of the lifecycle component that is being started. Adding a field to
describe the trigger that led to the component being invoked.

Starting with logging when an alarm caused a broadcast to start in the
app. Other examples of triggers that can be added are jobs for services,
push-messages for broadcasts, etc.

Test: make statsd_testdrive
statsd_testdrive 169

Bug: 242928495
Change-Id: Iffa03bd99f9ef6907444b3e3204d2a2f13e28102
Merged-In: Iffa03bd99f9ef6907444b3e3204d2a2f13e28102

Former-commit-id: 64d4919530e55f508e3a146fc7eaa13a2fa8e4f5