AppOps refactor for mode watchers.
The major work in this CL is moving op mode watchers to the
AppOpsServiceInterface and it's implementation.

Specifically this change introduces OnModeChangeListener interface that
contains the method that should be triggered on the mode change of the
op.
It also includes ModeChangedListenerDetails that encapsulates the
details about the listener.

In AppOpsService, ModeCallback class is now only a wrapper around the
oIBinder object.

BUG:240200048

Test: Manual
Change-Id: Iec633403367c1145727a641fb51ef1f3dc6cac41

Former-commit-id: 30342fdad7992356f657f01988243fb3d08cc8d3