Remove rounded corners on Tv PiP
The TV PiP has a border that has rounded corners. We don't want to have
rounded corners on the PiP surface, because it is an expensive
operation. Additionally, some TVs send video buffers directly to HWC, so
the buffers are not GPU-readable. This makes rounding corners in SF not
possible. To have a consistent PiP UI across ATV, we use the PiP border
as a means to add rounded corners.

Bug: 241766265
Test: m && flash && start pip
Change-Id: Ic45895a489cfce2c77398363b220802fdd52187b

Former-commit-id: eb180f4a8d5f0b09e58ad5537a45c80f01346999