Clean divider-remote-animating flag when exit split
The flag is used to control divider bar animation. We should also
clean the flag while exit split-screen to prevent split can not
update the divider bar visibility. Currently, the flag is clean
from remote animation finished or canceled. It’might be have the
possibility of not cleaning the flag if launch aborted.

Bug: 241861895
Test: manual
Test: pass existing tests
Change-Id: I07d1888957496dab8d5886eb7d9b21cc133b4214

Former-commit-id: 032fc54d6e7f5abec069b0b6ab4d280c4a39b61b