Include embedding activity change in transition
When enter ActivityEmbedding split, we will reparent the fill parent
activity to an embedded TaskFragment of different size. Before, because
the embedded TaskFragment is also a change target, the activity will be
excluded from the change targets. Now, we also include such activity so
that we can animate the transition of it.

Bug: 207070762
Test: atest WmTests:TransitionTests#testIncludeEmbeddedActivityReparent
Change-Id: I2de3a6e9ba2f74e455c56abd24010fff2c51e53d

Former-commit-id: 1ba311c20a421664f37f1b5f65fa1b887bf05884